refactor: cull thin polylogue modules
Collapse dead wrappers and tiny helper modules into their neighboring owners. Update imports and tests to match the leaner module layout, then verify with build and full pytest.
