+++ polylogue/pipeline/services/__init__.py
@@ -14,8 +14,7 @@ from __future__ import annotations
 
 from .acquisition import AcquireResult, AcquisitionService
 from .indexing import IndexService
-from .ingest_state import IngestPhase, IngestState
-from .parsing import ParseResult, ParsingService
+from .parsing import IngestPhase, IngestState, ParseResult, ParsingService
 from .planning import IngestPlan, PlanningService
 from .rendering import RenderResult, RenderService
 from .validation import ValidateResult, ValidationService
