+++ polylogue/pipeline/services/indexing.py
@@ -3,16 +3,10 @@
 from __future__ import annotations
 
 import sqlite3
-from collections.abc import AsyncIterable, Iterable
+from collections.abc import AsyncIterable, AsyncIterator, Iterable
 from typing import TYPE_CHECKING
 
 from polylogue.lib.log import get_logger
-from polylogue.storage.async_index import (
-    ensure_index,
-    index_status,
-    rebuild_index,
-    update_index_for_conversations,
-)
 
 if TYPE_CHECKING:
     from polylogue.config import Config
@@ -21,6 +15,108 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 __all__ = ["IndexService"]
+
+
+async def ensure_index(backend: SQLiteBackend) -> None:
+    """Create the FTS5 index table if it does not exist."""
+    async with backend.connection() as conn:
+        await conn.execute(
+            """
+            CREATE VIRTUAL TABLE IF NOT EXISTS messages_fts USING fts5(
+                message_id UNINDEXED,
+                conversation_id UNINDEXED,
+                text,
+                tokenize='unicode61'
+            );
+            """
+        )
+
+
+async def rebuild_index(backend: SQLiteBackend) -> None:
+    """Rebuild the entire FTS5 index from message rows."""
+    async with backend.connection() as conn:
+        await ensure_index(backend)
+        await conn.execute("DELETE FROM messages_fts")
+        await conn.execute(
+            """
+            INSERT INTO messages_fts (message_id, conversation_id, text)
+            SELECT messages.message_id, messages.conversation_id, messages.text
+            FROM messages
+            WHERE messages.text IS NOT NULL
+            """
+        )
+        await conn.commit()
+
+
+async def update_index_for_conversations(
+    conversation_ids: Iterable[str] | AsyncIterable[str],
+    backend: SQLiteBackend,
+) -> None:
+    """Update the FTS5 index for the provided conversations."""
+    async with backend.connection() as conn:
+        await ensure_index(backend)
+
+        async for chunk_ids in _chunked_ids(conversation_ids, size=500):
+            placeholders = ", ".join("?" for _ in chunk_ids)
+            params = tuple(chunk_ids)
+
+            await conn.execute(
+                f"DELETE FROM messages_fts WHERE conversation_id IN ({placeholders})",
+                params,
+            )
+            await conn.execute(
+                f"""
+                INSERT INTO messages_fts (message_id, conversation_id, text)
+                SELECT message_id, conversation_id, text
+                FROM messages
+                WHERE text IS NOT NULL AND conversation_id IN ({placeholders})
+                """,
+                params,
+            )
+
+        await conn.commit()
+
+
+async def _chunked_ids(
+    items: Iterable[str] | AsyncIterable[str],
+    *,
+    size: int,
+) -> AsyncIterator[list[str]]:
+    chunk: list[str] = []
+    async for item in _iter_ids(items):
+        chunk.append(item)
+        if len(chunk) >= size:
+            yield chunk
+            chunk = []
+
+    if chunk:
+        yield chunk
+
+
+async def _iter_ids(items: Iterable[str] | AsyncIterable[str]) -> AsyncIterator[str]:
+    if isinstance(items, AsyncIterable):
+        async for item in items:
+            yield item
+        return
+
+    for item in items:
+        yield item
+
+
+async def index_status(backend: SQLiteBackend) -> dict[str, object]:
+    """Return whether the FTS5 index exists and how many docs it contains."""
+    async with backend.connection() as conn:
+        cursor = await conn.execute(
+            "SELECT name FROM sqlite_master WHERE type='table' AND name='messages_fts'"
+        )
+        row = await cursor.fetchone()
+        exists = bool(row)
+        count = 0
+        if exists:
+            cursor = await conn.execute("SELECT COUNT(*) FROM messages_fts_docsize")
+            row = await cursor.fetchone()
+            count = row[0] if row else 0
+        return {"exists": exists, "count": int(count)}
 
 
 class IndexService:
