--- polylogue/showcase/__init__.py
@@ -1,7 +1,0 @@
-"""Showcase module: exercise all CLI commands and generate documentation artifacts."""
-
-from __future__ import annotations
-
-from polylogue.showcase.runner import ShowcaseRunner
-
-__all__ = ["ShowcaseRunner"]
