--- polylogue/sources/parsers/__init__.py
@@ -1,15 +1,0 @@
-"""Provider-specific parsers: JSON export → ParsedConversation."""
-
-from .base import (
-    ParsedAttachment,
-    ParsedConversation,
-    ParsedMessage,
-    RawConversationData,
-)
-
-__all__ = [
-    "ParsedAttachment",
-    "ParsedConversation",
-    "ParsedMessage",
-    "RawConversationData",
-]
