--- polylogue/sources/providers/__init__.py
@@ -1,15 +1,0 @@
-"""Provider-specific typed models for parse-don't-validate pattern."""
-
-from polylogue.sources.providers.chatgpt import ChatGPTConversation
-from polylogue.sources.providers.claude_ai import ClaudeAIConversation
-from polylogue.sources.providers.claude_code import ClaudeCodeRecord
-from polylogue.sources.providers.codex import CodexRecord
-from polylogue.sources.providers.gemini import GeminiMessage
-
-__all__ = [
-    "ChatGPTConversation",
-    "ClaudeAIConversation",
-    "ClaudeCodeRecord",
-    "CodexRecord",
-    "GeminiMessage",
-]
