--- polylogue/storage/async_index.py
@@ -1,149 +1,0 @@
-"""Async full-text search index management for SQLite.
-
-Provides async/await API for FTS5 index creation, rebuilding, and status checking.
-All operations use the SQLiteBackend for non-blocking database access.
-"""
-
-from __future__ import annotations
-
-from collections.abc import AsyncIterable, AsyncIterator, Iterable
-
-from polylogue.storage.backends.async_sqlite import SQLiteBackend
-
-
-async def ensure_index(backend: SQLiteBackend) -> None:
-    """Create FTS5 index table if it doesn't exist.
-
-    Args:
-        backend: Async SQLite backend instance
-    """
-    async with backend.connection() as conn:
-        await conn.execute(
-            """
-            CREATE VIRTUAL TABLE IF NOT EXISTS messages_fts USING fts5(
-                message_id UNINDEXED,
-                conversation_id UNINDEXED,
-                text,
-                tokenize='unicode61'
-            );
-            """
-        )
-
-
-async def rebuild_index(backend: SQLiteBackend) -> None:
-    """Rebuild the entire FTS5 search index from scratch.
-
-    Args:
-        backend: Async SQLite backend instance
-    """
-    async with backend.connection() as conn:
-        await ensure_index(backend)
-        await conn.execute("DELETE FROM messages_fts")
-        await conn.execute(
-            """
-            INSERT INTO messages_fts (message_id, conversation_id, text)
-            SELECT messages.message_id, messages.conversation_id, messages.text
-            FROM messages
-            WHERE messages.text IS NOT NULL
-            """
-        )
-        await conn.commit()
-
-
-async def update_index_for_conversations(
-    conversation_ids: Iterable[str] | AsyncIterable[str],
-    backend: SQLiteBackend,
-) -> None:
-    """Update FTS5 search index for specific conversations.
-
-    Optimized for batch operations using a single delete then batch insert.
-
-    Args:
-        conversation_ids: Conversation IDs to re-index
-        backend: Async SQLite backend instance
-    """
-    async with backend.connection() as conn:
-        await ensure_index(backend)
-
-        # Keep placeholder counts under SQLite parameter limits and avoid
-        # materializing all message rows in Python.
-        async for chunk_ids in _chunked_ids(conversation_ids, size=500):
-            placeholders = ", ".join("?" for _ in chunk_ids)
-            params = tuple(chunk_ids)
-
-            await conn.execute(
-                f"DELETE FROM messages_fts WHERE conversation_id IN ({placeholders})",
-                params,
-            )
-            await conn.execute(
-                f"""
-                INSERT INTO messages_fts (message_id, conversation_id, text)
-                SELECT message_id, conversation_id, text
-                FROM messages
-                WHERE text IS NOT NULL AND conversation_id IN ({placeholders})
-                """,
-                params,
-            )
-
-        await conn.commit()
-
-
-async def _chunked_ids(
-    items: Iterable[str] | AsyncIterable[str],
-    *,
-    size: int,
-) -> AsyncIterator[list[str]]:
-    """Split sync or async conversation IDs into bounded chunks."""
-    chunk: list[str] = []
-    async for item in _iter_ids(items):
-        chunk.append(item)
-        if len(chunk) >= size:
-            yield chunk
-            chunk = []
-
-    if chunk:
-        yield chunk
-
-
-async def _iter_ids(items: Iterable[str] | AsyncIterable[str]) -> AsyncIterator[str]:
-    """Normalize sync and async ID sources into one async iterator."""
-    if isinstance(items, AsyncIterable):
-        async for item in items:
-            yield item
-        return
-
-    for item in items:
-        yield item
-
-
-async def index_status(backend: SQLiteBackend) -> dict[str, object]:
-    """Get FTS5 index status information.
-
-    Args:
-        backend: Async SQLite backend instance
-
-    Returns:
-        Dictionary with 'exists' (bool) and 'count' (int) keys
-    """
-    async with backend.connection() as conn:
-        cursor = await conn.execute(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='messages_fts'"
-        )
-        row = await cursor.fetchone()
-        exists = bool(row)
-        count = 0
-        if exists:
-            # COUNT(*) on FTS virtual table is O(N) and extremely slow (minutes on large DBs).
-            # The backing docsize table has one row per indexed document and counts instantly.
-            cursor = await conn.execute("SELECT COUNT(*) FROM messages_fts_docsize")
-            row = await cursor.fetchone()
-            count = row[0] if row else 0
-        return {"exists": exists, "count": int(count)}
-
-
-__all__ = [
-    "ensure_index",
-    "rebuild_index",
-    "update_index_for_conversations",
-    "index_status",
-]
